chatbot/requirements.txt
langchain-community==0.2.3
langchain==0.2.3
langchain-google-genai==1.0.10
streamlit==1.35.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.29.2
faiss-cpu==1.7.4
python-dotenv

